Don't attempt to :chdir when the current buffer has no name, and therefore its directory is also empty
This doesn't actually do harm, but is not necessary.
Detect failure of fnamemodify() to convert an inexistent relative path to an absolute one, and attempt to convert the upwards ../../ path separately.
Add g:FilePathConvert_AdditionalIsFnamePattern to correctly grab URLs on Unix and Windows-style filespecs on Cygwin.
On Cygwin, also consider C:\-style filespecs as absolute paths.